20230201 20:35

+++ src/main/java/com/greenart/firstproject/config/security/JwtProperties.java
@@ -1,6 +1,6 @@
 package com.greenart.firstproject.config.security;
 
 public interface JwtProperties {
-    int EXPIRATION_TIME = 1000 * 60 * 30;
+    long EXPIRATION_TIME = 1000 * 60 * 30;
     String TOKEN_PREFIX = "Bearer ";
 }
